replace timezone with datetime
Since the plots are not really part of django they aren't returning timezone
objects which django can convert to the correct timezone. Therefore they are
always UTC. However we can disable the timezone support in django. The're might
be a more elegant solution to this but since this is just a small project I'll
leave it at that.